backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
psycopg2-binary
bcrypt
python-dotenv